Split fireStoreUtils.js into two files and change import pathing
