+++ src/components/profile/ConfigureProfile.js
@@ -9,7 +9,7 @@ import defaultAvatarSrc from '../../assets/profile-pic.png'
 import { useAuth } from '../../contexts/AuthContext'
 import { useDetails } from '../../contexts/UserDetailsContext'
 import { getDays, getMonths, getYears } from '../../utility/OptionSelectArrays'
-import { usernameAlreadyExists } from '../../utility/firestoreUtils'
+import { usernameAlreadyExists } from '../../utility/firestore/userDetailsFirestore'
 import Button from '../common/Button'
 import ImageInput from '../input/ImageInput'
 import LabelAndInput from '../input/LabelAndInput'
